Move Devtools helpers into lib.
