+++ packages/devtools/src/lib/singleton.ts
